requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.0
pydantic==2.5.0
pyyaml==6.6.0
